refactor(Application, Start) : Application클래스 분리
